requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
colorcet==3.1.0
datashader==0.18.1
idna==3.10
llvmlite==0.44.0
multipledispatch==1.0.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
param==2.2.1
pillow==11.3.0
pyct==0.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.16.0
six==1.17.0
toolz==1.0.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.7.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorcet==3.1.0
datashader==0.18.1
enum34==1.1.10
idna==3.10
llvmlite==0.44.0
multipledispatch==1.0.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
param==2.2.1
pillow==11.3.0
pyct==0.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.16.0
six==1.17.0
toolz==1.0.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.7.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorcet==3.1.0
datashader==0.18.1
dearpygui==2.1.0
enum34==1.1.10
idna==3.10
llvmlite==0.44.0
multipledispatch==1.0.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
param==2.2.1
pillow==11.3.0
pyct==0.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.16.0
six==1.17.0
toolz==1.0.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.7.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorcet==3.1.0
datashader==0.18.1
dearpygui==2.1.0
enum34==1.1.10
idna==3.10
llvmlite==0.44.0
multipledispatch==1.0.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
param==2.2.1
pillow==11.3.0
psycopg2==2.9.10
pyct==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scipy==1.16.0
six==1.17.0
toolz==1.0.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.7.1
